Removed htslib submodule
